TAS-95:(現在の学習進歩修正完了)
